Enabled external parameter sharing
Added functionality to save dictionaries of simulation features using pickle, so you can use the same SN over multiple simulations.  This also greatly cleans up the arguments needed to run the wrapper code.  Only thing to be careful of is pointing to the correct (previous) folder when making a new sim.